Move getTopicDensity and topicCount to the abstract researcher
