+++ packages/yoastseo/src/languageProcessing/AbstractResearcher.js
@@ -23,6 +23,8 @@ import readingTime from "./researches/readingTime";
 import h1s from "./researches/h1s";
 import sentences from "./researches/sentences";
 import findKeywordInFirstParagraph from "./researches/findKeywordInFirstParagraph.js";
+import getTopicDensity from "./researches/getTopicDensity";
+import topicCount from "./researches/topicCount";
 
 /**
  * The researches contains all the researches
@@ -56,6 +58,8 @@ export default class AbstractResearcher {
 			h1s: h1s,
 			sentences: sentences,
 			firstParagraph: findKeywordInFirstParagraph,
+			getTopicDensity: getTopicDensity,
+			topicCount: topicCount
 		};
 
 		this._data = {};
