+++ packages/yoastseo/src/languageProcessing/languages/en/Researcher.js
@@ -3,8 +3,8 @@ import keywordCount from "../../researches/_todo/keywordCount";
 import metaDescriptionKeyword from "../../researches/_todo/metaDescriptionKeyword.js";
 import keywordCountInUrl from "../../researches/_todo/keywordCountInUrl";
 import relevantWords from "../../researches/_todo/relevantWords";
-import getTopicDensity from "../../researches/_todo/getTopicDensity";
-import topicCount from "../../researches/_todo/topicCount";
+import getTopicDensity from "../../researches/getTopicDensity";
+import topicCount from "../../researches/topicCount";
 import { keyphraseDistributionResearcher } from "../../researches/_todo/keyphraseDistribution";
 import getProminentWordsForInsights from "../../researches/_todo/getProminentWordsForInsights";
 import getProminentWordsForInternalLinking from "../../researches/_todo/getProminentWordsForInternalLinking";
